# swagger.json
{"openapi":"3.0.0","paths":{"/User":{"get":{"operationId":"AuthController_getAllPeople","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}}},"post":{"operationId":"AuthController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"The user has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"403":{"description":"Forbidden."},"409":{"description":"User with the same ID already exists"}}}},"/User/{id}":{"get":{"operationId":"AuthController_getUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"404":{"description":"User not found"}}},"patch":{"operationId":"AuthController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"404":{"description":"User not found"}}},"delete":{"operationId":"AuthController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}},"404":{"description":"User not found"}}}}},"info":{"title":"Median","description":"The Median API description","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UserEntity":{"type":"object","properties":{"id":{"type":"number"},"firstname":{"type":"string"},"lastname":{"type":"string"},"age":{"type":"number"}},"required":["id","firstname","lastname","age"]},"CreateUserDto":{"type":"object","properties":{"firstname":{"type":"string","description":"First name","example":"John"},"lastname":{"type":"string","description":"Last name","example":"Doe"},"age":{"type":"number","description":"Age","example":30}},"required":["firstname","lastname","age"]},"UpdateUserDto":{"type":"object","properties":{"firstname":{"type":"string","description":"First name","example":"John"},"lastname":{"type":"string","description":"Last name","example":"Doe"},"age":{"type":"number","description":"Age","example":30}}}}}}